Khanh dev

+++ smart-watering-backend/src/main/java/com/iot/smartwatering/smart_watering_backend/repository/SensorDataRepository.java
@@ -12,18 +12,18 @@ import java.util.List;
 
 @Repository
 public interface SensorDataRepository extends JpaRepository<SensorData, Long> {
-    List<SensorData> findByZone_ZoneIdOrderByCreatedAtDesc(Integer zoneId);
+    List<SensorData> findByZone_ZoneIdOrderByCreatedAtDesc(Long zoneId);
 
     @Query("SELECT sd FROM SensorData sd WHERE sd.zone.zoneId = :zoneId " +
             "AND sd.createdAt BETWEEN :startDate AND :endDate " +
             "ORDER BY sd.createdAt DESC")
     List<SensorData> findByZoneAndDateRange(
-            @Param("zoneId") Integer zoneId,
+            @Param("zoneId") Long zoneId,
             @Param("startDate") LocalDateTime startDate,
             @Param("endDate") LocalDateTime endDate
     );
 
     @Query("SELECT sd FROM SensorData sd WHERE sd.zone.zoneId = :zoneId " +
             "ORDER BY sd.createdAt DESC LIMIT 1")
-    SensorData findLatestByZone(@Param("zoneId") Integer zoneId);
+    SensorData findLatestByZone(@Param("zoneId") Long zoneId);
 }
